fix: remove tiyins - all monetary values are whole UZS (sum)
- Add Currency Convention section to frontend spec (formatMoney utility,
  money fields catalog, anti-patterns guide)
- Update spec v3.0.0 -> v3.1.0 with breaking change #11 (no tiyins)
- Fix all code descriptions: 'Тийины/тийинах/копейки' -> 'сум (UZS)'
- Fix total_amount: float -> int in application/order/schemas.py
- Update contract model comments: 'в тийинах' -> 'в сумах (UZS)'
- All TypeScript types now annotated with '// сум (UZS)' on money fields
- Test checklist updated: display without decimal places

+++ src/application/order/schemas.py
@@ -106,6 +106,6 @@ class OrderResponse(BaseModel):
     client_inventory_id: uuid.UUID
     payment_method: PaymentMethod
     status: str
-    total_amount: float  # Или Decimal
+    total_amount: int
 
     model_config = ConfigDict(from_attributes=True)
